requirements.txt
fusepy==3.0.1
requests==2.32.5
plyvel==1.5.1
python-dateutil==2.9.0.post0
google-api-python-client==2.187.0
google-auth-oauthlib==1.2.3
gitignore_parser==0.1.13
humanize==4.15.0
